Adding Unity
This commit is required for me to finish setting up GitHub and Unity Intergration. It involved me creating a Github Repository and a new unity project. I saved the repository to my desktop and then copied and pasted all the asset files over from the unity project to the Github Repository File and finally opening up the File as a unity project.