Player Movement Script
This commit adds in the player movement script. The way I have done it is by adding a Rigidbody2D to a sprite and then transforming the sprite along a vector that I make based on wether or not an input is detected from the WASD keys or from the arrow keys. When an input is detected, the input value is timsed by the speed variable which is public, allowing me to easily modify the player speed as I go on if I feel it needs balancing.