Main Menu
This commit adds in the Main menu with both the quit and start button that is able to take you to the first level scene. Right now it just takes you to what will be the intro scene as a test but will be fixed once the first level is made.